Fix the bugs so that the tests pass

+++ src/cellkit/model/norm.py
@@ -11,4 +11,4 @@ class RMSNorm(torch.nn.Module):
     
     def forward(self, x):
         output = self._norm(x.float()).type_as(x)
-        return output * self.weight
+        return output * self.weight.to(dtype=output.dtype)
